fetch identities outside of react
